LPFE-88: Moved line-item.service to models
